refactor: artifact checksum
