+++ internal/pipeline/artifactory/artifactory.go
@@ -52,7 +52,6 @@ func (Pipe) Default(ctx *context.Context) error {
 //
 // Docs: https://www.jfrog.com/confluence/display/RTF/Artifactory+REST+API#ArtifactoryRESTAPI-Example-DeployinganArtifact
 func (Pipe) Run(ctx *context.Context) error {
-
 	if len(ctx.Config.Artifactories) == 0 {
 		return pipeline.Skip("artifactory section is not configured")
 	}
